+++ src/main/scala/mllab/evaluaton/Evaluation.scala
@@ -34,14 +34,14 @@ object Evaluation{
     println("F " + falsePositives + " " + falseNegatives)
   }
 
-  def MSE(y_pred: List[Float], y_true: List[Float]): Float =
-    (y_pred zip y_true).map{case (x, y) => Math.pow(x - y, 2)}.sum.toFloat
+  def MSE(y_pred: List[Double], y_true: List[Double]): Double =
+    (y_pred zip y_true).map{case (x, y) => Math.pow(x - y, 2)}.sum
 
   /*
   Mean squared error scaled
   */
-  def MSES(y_pred: List[Float], y_true: List[Float]): Float = {
+  def MSES(y_pred: List[Double], y_true: List[Double]): Double = {
     val MSEval = (y_pred zip y_true).map{case (x, y) => Math.pow(x - y, 2)}.sum
-    (MSEval / Math.abs(y_true.sum)).toFloat
+    MSEval / Math.abs(y_true.sum)
   }
 }
